+++ testutils/src/main/java/com/rokt/core/testutils/BaseViewModelTest.kt
@@ -1,4 +1,4 @@
-package com.core.testutils
+package com.rokt.core.testutils
 
 import org.junit.Rule