+++ testutils/src/main/java/com/rokt/core/testutils/MockkUnitTest.kt
@@ -1,4 +1,4 @@
-package com.core.testutils
+package com.rokt.core.testutils
 
 import io.mockk.clearAllMocks
 import io.mockk.unmockkAll