+++ testutils/src/main/java/com/rokt/core/testutils/annotations/DcuiOfferJson.kt
@@ -1,4 +1,4 @@
-package com.core.testutils.annotations
+package com.rokt.core.testutils.annotations
 
 @Target(AnnotationTarget.FUNCTION)
 annotation class DcuiOfferJson(val jsonString: String = "", val jsonFile: String = "")