+++ testutils/src/main/java/com/rokt/core/testutils/assertion/SpacingAssertions.kt
@@ -1,4 +1,4 @@
-package com.core.testutils.assertion
+package com.rokt.core.testutils.assertion
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.offset