+++ testutils/src/main/java/com/rokt/core/testutils/mockserver/CapturedEvent.kt
@@ -1,3 +1,3 @@
-package com.core.testutils.mockserver
+package com.rokt.core.testutils.mockserver
 
 data class CapturedEvent(val sessionId: String, val eventType: String, val parentGuid: String, val token: String)