+++ testutils/src/main/java/com/rokt/core/testutils/rule/ScreenshotConfig.kt
@@ -1,4 +1,4 @@
-package com.core.testutils.rule
+package com.rokt.core.testutils.rule
 
 @Target(AnnotationTarget.FUNCTION)
 annotation class ScreenshotConfig(val captureOnSuccess: Boolean = false)